requirements.txt
# aiobotocore>=2.9.0
aiohttp==3.9.1
aioitertools==0.11.0
aiosignal==1.3.1
asttokens==2.4.1
attrs==23.1.0
# boto3>=1.34.7
# botocore>=1.34.9
certifi==2023.11.17
charset-normalizer==3.3.2
colorama==0.4.6
comm==0.2.0
cramjam==2.7.0
# debugpy==1.8.0
decorator==5.1.1
executing==2.0.1
fastparquet==2023.10.1
frozenlist==1.4.1
fsspec==2023.12.2
idna==3.6
# ipykernel==6.26.0
# ipython==8.17.2
jedi==0.19.1
jmespath==1.0.1
jupyter_client==8.6.0
jupyter_core==5.5.0
matplotlib-inline==0.1.6
multidict==6.0.4
nest-asyncio==1.5.8
numpy==1.26.2
packaging==23.2
pandas==2.1.4
parso==0.8.3
platformdirs==3.11.0
prompt-toolkit==3.0.39
psutil==5.9.6
pure-eval==0.2.2
# pyarrow==14.0.2
Pygments==2.16.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
pywin32==306;sys_platform == 'win32'
pyzmq==25.1.1
requests==2.31.0
s3fs==2023.12.2
s3transfer==0.10.0
six==1.16.0
stack-data==0.6.3
tornado==6.3.3
traitlets==5.13.0
tzdata==2023.3
urllib3==2.0.7
wcwidth==0.2.9
wrapt==1.16.0
yarl==1.9.4
ta==0.11.0
